tournament commands to team command
