+++ config/snowflakes-commands-example.json
@@ -12,7 +12,7 @@
     "slashRank": "",
     "slashRole": "",
     "slashTeam": "",
-    "slashTournament": "",
+    "slashTournaments": "",
     "slashUser": "",
     "slashVoice": "",
     "userModeration": ""
