+++ registry/slashTeam.js
@@ -1,14 +1,25 @@
 // @ts-check
 const u = require("./regUtils");
 
-// ROLE COMMANDS
-
+// GENERAL FUNCTIONS
 const role = (action = "add") => new u.string()
   .setName(action === 'equip' ? "color" : "role")
   .setDescription(`The role to ${action}`)
   .setRequired(action === 'equip' ? false : true)
   .setAutocomplete(true);
+/**
+ *
+ * @param {number | null} num
+ * @param {boolean} [req]
+ * @returns
+ */
+const user = (num, req = false) => new u.user()
+  .setName(`${num ?? "user"}`)
+  .setDescription(`User ${num ?? ""}`)
+  .setRequired(num === 1 || req);
 
+
+// ROLE COMMANDS
 const give = new u.sub()
   .setName("give")
   .setDescription("Give someone a role")
@@ -33,7 +44,6 @@ const take = new u.sub()
 
 
 // BANK COMMANDS
-
 const award = new u.sub()
   .setName("award")
   .setDescription("Award ember to a member for the house cup.")
@@ -58,6 +68,28 @@ const award = new u.sub()
       .setRequired(false)
   );
 
+
+// TOURNAMENT COMMANDS
+const champion = new u.sub()
+  .setName("champions")
+  .setDescription("Declare tournament champions!")
+  .addStringOption(
+    new u.string()
+      .setName("tournament")
+      .setDescription("The name of the tournament")
+      .setRequired(true)
+  )
+  .addUserOption(user(1))
+  .addUserOption(user(2))
+  .addUserOption(user(3))
+  .addUserOption(user(4))
+  .addUserOption(user(5))
+  .addUserOption(user(6));
+
+const reset = new u.sub()
+  .setName("reset")
+  .setDescription("Reset the Tournament Participant role");
+
 module.exports = new u.cmd()
   .setName("team")
   .setDescription("Do team stuff. idk.")
@@ -73,6 +105,13 @@ module.exports = new u.cmd()
       .setName("bank")
       .setDescription("Interact with currency")
       .addSubcommand(award)
+  )
+  .addSubcommandGroup(
+    new u.subGroup()
+      .setName("tournament")
+      .setDescription("Manage tournaments")
+      .addSubcommand(champion)
+      .addSubcommand(reset)
   )
   .setDMPermission(false)
   .setDefaultMemberPermissions(u.privateCommand)
